requirements.txt
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
asyncio-mqtt==0.16.2
asyncpg==0.30.0
bcrypt==4.1.3
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.5
ecdsa==0.19.1
fastapi==0.116.1
gevent==25.5.1
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
logzero==1.7.0
Mako==1.3.10
MarkupSafe==3.0.2
numpy==2.3.2
packaging==25.0
paho-mqtt==2.1.0
pandas==2.3.1
passlib==1.7.4
pluggy==1.6.0
prometheus_client==0.22.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pyotp==2.9.0
pytest==8.4.1
pytest-asyncio==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
redis==6.2.0
requests==2.32.4
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
smartapi-python==1.5.5
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.47.2
structlog==25.4.0
ta==0.11.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
websocket==0.2.1
websocket-client==1.8.0
websockets==15.0.1
zope.event==5.1.1
zope.interface==7.2
